shift quotes logic
