+++ src/App.css
@@ -48,6 +48,7 @@
   justify-content: center;
   align-items: center;
   display: flex;
+  flex-direction: column;
 }
 
 .main-wrapper {
